Change over to use protobufjs
Using protobufjs instead of google-protobuf because it is ~ 1/10 the size of the google library. Using the google library added >100kb to bundle size and it could not be split or tree shaken in an efficient way. This caused the bundle corresponding to the api to be larger than 200kb value we are targeting. Not to mention protobufjs is much more widely used according to npm